regions plugin: move classes to separate files (#1934)
* - splitted regions.js in two files, one for RegionsPlugin class, one for Region class, and moved both to plugin/regions dir
- added findInDirectory function in the plugin build file, so the plugins entries can be found also if in deeper directories, recursively

* added changelog and versioning

* - renamed regions.js to index.js
- updated the way of building plugins, in particular using parent directory name for the plugin when its name is index.js
- updated versioning and comments

* fixed merge mismatch in regions/index.js when instantiate Region class
